snc/requirements-snc.txt
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
celery[redis]==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==37.0.1
dill==0.3.5.1
idna==3.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
neo4j==4.4.4
numexpr==2.8.3
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pick==2.0.0
pip==21.2.4
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
setuptools==61.2.0
six==1.16.0
tweepy==4.10.0
urllib3==1.26.9
wheel==0.37.1
